refactoring, move to a single file project structure
